+++ libs/minimizerscpp/src/kthread.cpp
@@ -7,8 +7,8 @@
 #ifdef USE_STL
 #include <condition_variable>
 #include <mutex>
-#include <vector>
 #include <thread>
+#include <vector>
 #else
 #include <pthread.h>
 #endif
@@ -66,14 +66,14 @@ static void* ktf_worker( void* data )
             break;
 #ifdef USE_STL
         // dirty ...
-        w->t->func( w->t->data, i, w - &( w->t->w[ 0 ] ) );
+        w->t->func( w->t->data, i, (int)( w - &( w->t->w[ 0 ] ) ) );
 #else
         w->t->func( w->t->data, i, w - w->t->w );
 #endif
     }
     while( ( i = steal_work( w->t ) ) >= 0 )
 #ifdef USE_STL
-        w->t->func( w->t->data, i, w - &( w->t->w[ 0 ] ) );
+        w->t->func( w->t->data, i, (int)( w - &( w->t->w[ 0 ] ) ) );
 #else
         w->t->func( w->t->data, i, w - w->t->w );
 
@@ -223,9 +223,9 @@ static void* ktp_worker( void* data )
         } // scope lock_guard
     } // while
 #ifdef USE_STL
-	return 0;
+    return 0;
 #else
-	pthread_exit(0);
+    pthread_exit( 0 );
 #endif
 } // function
